OnlinemarketProjectWebParent/OnlinemarketProjectFrontend/src/main/java/com/onlinemarket/OnlinemarketProjectFrontend/setting/SettingRepository.java
package com.onlinemarket.OnlinemarketProjectFrontend.setting;

import java.util.List;

import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import com.onlinemarket.OnlinemarketProjectCommon.entity.Setting;
import com.onlinemarket.OnlinemarketProjectCommon.entity.SettingCategory;

public interface SettingRepository extends CrudRepository<Setting, String>{

    @Query("SELECT s FROM Setting s WHERE s.category = ?1 OR s.category = ?2")
    public List<Setting> findByTwoCategories(SettingCategory catOne, SettingCategory catTwo);

    public List<Setting> findByCategory(SettingCategory category);

}
